:sparkles: Divide modules
