+++ ProcessingKit/Core/Constants/Constants.swift
@@ -21,25 +21,3 @@ public protocol Constants {
     var TAU: CGFloat { get }
     var TWO_PI: CGFloat { get }
 }
-
-extension ProcessingView: Constants {
-    public var HALF_PI: CGFloat {
-        return .pi / 2
-    }
-
-    public var PI: CGFloat {
-        return .pi
-    }
-
-    public var QUARTER_PI: CGFloat {
-        return .pi / 4
-    }
-
-    public var TAU: CGFloat {
-        return self.TWO_PI
-    }
-
-    public var TWO_PI: CGFloat {
-        return .pi * 2
-    }
-}
