+++ ProcessingKit/Core/Extensions/NSImage.swift
