+++ ProcessingKit/Core/Extensions/String.swift
