+++ ProcessingKit/Core/Structure/Loop.swift
@@ -1,0 +1,14 @@
+//
+//  Structure.swift
+//  ProcessingKit
+//
+//  Created by AtsuyaSato on 2017/08/13.
+//  Copyright © 2017年 Atsuya Sato. All rights reserved.
+//
+
+import Foundation
+
+public protocol LoopModelContract {
+    func loop()
+    func noLoop()
+}
